muchos cambios

+++ src/main/java/com/Tienda/Franco/Repository/ProductoRepository.java
@@ -1,10 +1,11 @@
 package com.Tienda.Franco.Repository;
 
-import com.Tienda.Franco.Model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.Tienda.Franco.Model.Producto;
+
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 
 }
